Remove freezable implementation directory
