+++ packages/lib/test/src/directory/ImplementationDirectory.test.js
@@ -2,8 +2,50 @@
 require('../../setup')
 
 import ImplementationDirectory from '../../../src/directory/ImplementationDirectory'
-import shouldBehaveLikeImplementationDirectory from './BaseImplementationDirectory.behavior';
+import Contracts from '../../../src/utils/Contracts'
+
+const DummyImplementationV2 = Contracts.getFromLocal('DummyImplementationV2')
 
 contract('ImplementationDirectory', function(accounts) {
-  shouldBehaveLikeImplementationDirectory(ImplementationDirectory, accounts)
+  const [_, owner] = accounts
+  const txParams = { from: owner }
+
+  beforeEach('deploying implementation directory', async function () {
+    this.directory = await ImplementationDirectory.deploy(txParams)
+  })
+
+  it('has an address', async function () {
+    (await this.directory.address).should.not.be.null
+  })
+
+  it('has an owner', async function () {
+    (await this.directory.owner()).should.be.equal(owner)
+  })
+
+  it('can set new implementations', async function () {
+    const implementation = await DummyImplementationV2.new()
+    await this.directory.setImplementation('DummyImplementation', implementation.address)
+
+    const currentImplementation = await this.directory.getImplementation('DummyImplementation');
+    currentImplementation.should.be.eq(implementation.address)
+  })
+
+  it('can unset implementations', async function () {
+    const implementation = await DummyImplementationV2.new()
+    await this.directory.setImplementation('DummyImplementation', implementation.address)
+    await this.directory.unsetImplementation('DummyImplementation')
+
+    const currentImplementation = await this.directory.getImplementation('DummyImplementation')
+    currentImplementation.should.be.zeroAddress
+  }) 
+
+  it('can be frozen', async function () {
+    let frozen = await this.directory.isFrozen();
+    frozen.should.be.false
+
+    await this.directory.freeze().should.eventually.be.fulfilled
+
+    frozen = await this.directory.isFrozen()
+    frozen.should.be.true
+  })
 })
